refactor(executor): Split grouping_sets.rs into smaller modules
Split the 1561-line grouping_sets.rs file into logical modules for
better maintainability and navigation:

- mod.rs (190 lines): Core types (ResolvedGroupingSet, GroupingContext)
  and public API re-exports
- expansion.rs (467 lines): ROLLUP, CUBE, and GROUPING SETS expansion
  logic with expand_* functions
- expression_utils.rs (715 lines): Expression equality comparison for
  GROUPING() function matching
- alias_resolution.rs (254 lines): GROUP BY alias and position resolution

Note: Two files exceed the suggested 400-line limit due to comprehensive
test coverage (~437 and ~189 lines of tests respectively). The actual
functional code in each file is well under 400 lines (~277 lines each).

- All existing tests pass
- No public API changes (same exports as before)
- No functional changes

Closes #3053
